refactor: remove workflow specific form fields and make form fields able to take a variable picker
