+++ packages/twenty-front/src/modules/object-record/record-field/components/FormFieldInput.tsx
@@ -2,6 +2,7 @@ import { FieldMetadataItem } from '@/object-metadata/types/FieldMetadataItem';
 import { FormBooleanFieldInput } from '@/object-record/record-field/form-types/components/FormBooleanFieldInput';
 import { FormNumberFieldInput } from '@/object-record/record-field/form-types/components/FormNumberFieldInput';
 import { FormTextFieldInput } from '@/object-record/record-field/form-types/components/FormTextFieldInput';
+import { VariablePickerComponent } from '@/object-record/record-field/form-types/types/VariablePickerComponent';
 import { isFieldBoolean } from '@/object-record/record-field/types/guards/isFieldBoolean';
 import { isFieldNumber } from '@/object-record/record-field/types/guards/isFieldNumber';
 import { isFieldText } from '@/object-record/record-field/types/guards/isFieldText';
@@ -11,12 +12,14 @@ type FormFieldInputProps = {
   field: FieldMetadataItem;
   defaultValue: JsonValue;
   onPersist: (value: JsonValue) => void;
+  VariablePicker?: VariablePickerComponent;
 };
 
 export const FormFieldInput = ({
   field,
   defaultValue,
   onPersist,
+  VariablePicker,
 }: FormFieldInputProps) => {
   return isFieldNumber(field) ? (
     <FormNumberFieldInput
@@ -25,6 +28,7 @@ export const FormFieldInput = ({
       defaultValue={defaultValue as string | number | undefined}
       onPersist={onPersist}
       placeholder={field.label}
+      VariablePicker={VariablePicker}
     />
   ) : isFieldBoolean(field) ? (
     <FormBooleanFieldInput
@@ -32,6 +36,7 @@ export const FormFieldInput = ({
       label={field.label}
       defaultValue={defaultValue as string | boolean | undefined}
       onPersist={onPersist}
+      VariablePicker={VariablePicker}
     />
   ) : isFieldText(field) ? (
     <FormTextFieldInput
@@ -39,6 +44,7 @@ export const FormFieldInput = ({
       defaultValue={defaultValue as string | undefined}
       onPersist={onPersist}
       placeholder={field.label}
+      VariablePicker={VariablePicker}
     />
   ) : null;
 };
