+++ packages/twenty-front/src/modules/object-record/record-field/form-types/components/FormBooleanFieldInput.tsx
@@ -1,16 +1,30 @@
-import { FormFieldInputBase } from '@/object-record/record-field/form-types/components/FormFieldInputBase';
+import {
+  StyledContainer,
+  StyledInputContainer,
+  StyledRowContainer,
+} from '@/object-record/record-field/form-types/components/FormFieldInputBase';
+import { VariablePickerComponent } from '@/object-record/record-field/form-types/types/VariablePickerComponent';
 import { BooleanInput } from '@/ui/field/input/components/BooleanInput';
+import { InputLabel } from '@/ui/input/components/InputLabel';
+import { SortOrFilterChip } from '@/views/components/SortOrFilterChip';
+import { StyledVariableContainer } from '@/workflow/components/WorkflowFormFieldInputBase';
+import { extractVariableLabel } from '@/workflow/search-variables/utils/extractVariableLabel';
+import { isStandaloneVariableString } from '@/workflow/utils/isStandaloneVariableString';
 import styled from '@emotion/styled';
-import { useState } from 'react';
+import { useId, useState } from 'react';
+import { isDefined } from 'twenty-ui';
 
 const StyledBooleanInputContainer = styled.div`
   padding-inline: ${({ theme }) => theme.spacing(2)};
 `;
 
+type EditingMode = 'input' | 'variable';
+
 type FormBooleanFieldInputProps = {
   label?: string;
   defaultValue: boolean | string | undefined;
   onPersist: (value: boolean | null | string) => void;
+  VariablePicker?: VariablePickerComponent;
   readonly?: boolean;
 };
 
@@ -19,7 +33,16 @@ export const FormBooleanFieldInput = ({
   defaultValue,
   onPersist,
   readonly,
+  VariablePicker,
 }: FormBooleanFieldInputProps) => {
+  const inputId = useId();
+
+  const defaultEditingMode = isStandaloneVariableString(defaultValue)
+    ? 'variable'
+    : 'input';
+  const [editingMode, setEditingMode] =
+    useState<EditingMode>(defaultEditingMode);
+
   const [draftValue, setDraftValue] = useState(defaultValue ?? false);
 
   const handleChange = (newValue: boolean) => {
@@ -28,20 +51,51 @@ export const FormBooleanFieldInput = ({
     onPersist(newValue);
   };
 
+  const handleVariableTagInsert = (variable: string) => {
+    setEditingMode('variable');
+    setDraftValue(variable);
+
+    onPersist(variable);
+  };
+
+  const handleUnlinkVariable = () => {
+    setEditingMode('input');
+    setDraftValue(false);
+
+    onPersist(false);
+  };
+
   return (
-    <FormFieldInputBase
-      label={label}
-      Input={
-        <StyledBooleanInputContainer>
-          <BooleanInput
-            value={draftValue as boolean}
-            readonly={readonly}
-            onToggle={(newValue) => {
-              handleChange(newValue);
-            }}
+    <StyledContainer>
+      {label ? <InputLabel>{label}</InputLabel> : null}
+
+      <StyledRowContainer>
+        <StyledInputContainer hasRightElement={isDefined(VariablePicker)}>
+          {editingMode === 'input' ? (
+            <StyledBooleanInputContainer>
+              <BooleanInput
+                value={draftValue as boolean}
+                readonly={readonly}
+                onToggle={handleChange}
+              />
+            </StyledBooleanInputContainer>
+          ) : (
+            <StyledVariableContainer>
+              <SortOrFilterChip
+                labelValue={extractVariableLabel(draftValue as string)}
+                onRemove={handleUnlinkVariable}
+              />
+            </StyledVariableContainer>
+          )}
+        </StyledInputContainer>
+
+        {VariablePicker ? (
+          <VariablePicker
+            inputId={inputId}
+            onVariableSelect={handleVariableTagInsert}
           />
-        </StyledBooleanInputContainer>
-      }
-    />
+        ) : null}
+      </StyledRowContainer>
+    </StyledContainer>
   );
 };
