+++ packages/twenty-front/src/modules/object-record/record-field/form-types/components/FormNumberFieldInput.tsx
@@ -1,7 +1,18 @@
-import { FormFieldInputBase } from '@/object-record/record-field/form-types/components/FormFieldInputBase';
+import {
+  StyledContainer,
+  StyledInputContainer,
+  StyledRowContainer,
+} from '@/object-record/record-field/form-types/components/FormFieldInputBase';
+import { VariablePickerComponent } from '@/object-record/record-field/form-types/types/VariablePickerComponent';
 import { TextInput } from '@/ui/field/input/components/TextInput';
+import { InputLabel } from '@/ui/input/components/InputLabel';
+import { SortOrFilterChip } from '@/views/components/SortOrFilterChip';
+import { StyledVariableContainer } from '@/workflow/components/WorkflowFormFieldInputBase';
+import { extractVariableLabel } from '@/workflow/search-variables/utils/extractVariableLabel';
+import { isStandaloneVariableString } from '@/workflow/utils/isStandaloneVariableString';
 import styled from '@emotion/styled';
-import { useState } from 'react';
+import { useId, useState } from 'react';
+import { isDefined } from 'twenty-ui';
 import {
   canBeCastAsNumberOrNull,
   castAsNumberOrNull,
@@ -11,11 +22,14 @@ const StyledInput = styled(TextInput)`
   padding: ${({ theme }) => `${theme.spacing(1)} ${theme.spacing(2)}`};
 `;
 
+type EditingMode = 'input' | 'variable';
+
 type FormNumberFieldInputProps = {
   label?: string;
   placeholder: string;
   defaultValue: number | string | undefined;
   onPersist: (value: number | null | string) => void;
+  VariablePicker?: VariablePickerComponent;
 };
 
 export const FormNumberFieldInput = ({
@@ -23,8 +37,17 @@ export const FormNumberFieldInput = ({
   placeholder,
   defaultValue,
   onPersist,
+  VariablePicker,
 }: FormNumberFieldInputProps) => {
+  const inputId = useId();
+
   const [draftValue, setDraftValue] = useState(defaultValue ?? '');
+
+  const defaultEditingMode = isStandaloneVariableString(defaultValue)
+    ? 'variable'
+    : 'input';
+  const [editingMode, setEditingMode] =
+    useState<EditingMode>(defaultEditingMode);
 
   const persistNumber = (newValue: string) => {
     if (!canBeCastAsNumberOrNull(newValue)) {
@@ -42,20 +65,52 @@ export const FormNumberFieldInput = ({
     persistNumber(newText.trim());
   };
 
+  const handleUnlinkVariable = () => {
+    setEditingMode('input');
+    setDraftValue('');
+
+    onPersist(null);
+  };
+
+  const handleVariableTagInsert = (variable: string) => {
+    setEditingMode('variable');
+    setDraftValue(variable);
+
+    onPersist(variable);
+  };
+
   return (
-    <FormFieldInputBase
-      label={label}
-      Input={
-        <StyledInput
-          placeholder={placeholder}
-          value={String(draftValue)}
-          copyButton={false}
-          hotkeyScope="record-create"
-          onChange={(value) => {
-            handleChange(value);
-          }}
-        />
-      }
-    />
+    <StyledContainer>
+      {label ? <InputLabel htmlFor={inputId}>{label}</InputLabel> : null}
+
+      <StyledRowContainer>
+        <StyledInputContainer hasRightElement={isDefined(VariablePicker)}>
+          {editingMode === 'input' ? (
+            <StyledInput
+              inputId={inputId}
+              placeholder={placeholder}
+              value={String(draftValue)}
+              copyButton={false}
+              hotkeyScope="record-create"
+              onChange={handleChange}
+            />
+          ) : (
+            <StyledVariableContainer>
+              <SortOrFilterChip
+                labelValue={extractVariableLabel(draftValue as string)}
+                onRemove={handleUnlinkVariable}
+              />
+            </StyledVariableContainer>
+          )}
+        </StyledInputContainer>
+
+        {VariablePicker ? (
+          <VariablePicker
+            inputId={inputId}
+            onVariableSelect={handleVariableTagInsert}
+          />
+        ) : null}
+      </StyledRowContainer>
+    </StyledContainer>
   );
 };
