+++ packages/twenty-front/src/modules/object-record/record-field/form-types/types/VariablePickerComponent.ts
@@ -1,0 +1,6 @@
+export type VariablePickerComponent = React.FC<{
+  inputId: string;
+  disabled?: boolean;
+  multiline?: boolean;
+  onVariableSelect: (variable: string) => void;
+}>;
