+++ packages/twenty-front/src/modules/workflow/components/WorkflowEditActionFormRecordCreate.tsx
@@ -1,7 +1,8 @@
 import { useFilteredObjectMetadataItems } from '@/object-metadata/hooks/useFilteredObjectMetadataItems';
+import { FormFieldInput } from '@/object-record/record-field/components/FormFieldInput';
 import { Select, SelectOption } from '@/ui/input/components/Select';
 import { WorkflowEditGenericFormBase } from '@/workflow/components/WorkflowEditGenericFormBase';
-import { WorkflowFormFieldInput } from '@/workflow/components/WorkflowFormFieldInput';
+import { WorkflowVariablePicker } from '@/workflow/components/WorkflowVariablePicker';
 import { WorkflowRecordCreateAction } from '@/workflow/types/Workflow';
 import { useTheme } from '@emotion/react';
 import { useEffect, useState } from 'react';
@@ -168,13 +169,14 @@ export const WorkflowEditActionFormRecordCreate = ({
         const currentValue = formData[field.name] as JsonValue;
 
         return (
-          <WorkflowFormFieldInput
+          <FormFieldInput
             key={field.id}
             defaultValue={currentValue}
             field={field}
             onPersist={(value) => {
               handleFieldChange(field.name, value);
             }}
+            VariablePicker={WorkflowVariablePicker}
           />
         );
       })}
